requirements.txt
imageio==2.25.1
matplotlib==3.5.3
numpy==1.21.6
opencv-python==4.6.0.66
pandas==1.3.5
Pillow==9.3.0
scikit-image==0.19.3
scipy==1.7.3
torch==1.7.1+cu110
torchvision==0.8.2+cu110